requirements.txt
# Core dependencies for Tennis Discovery Agent

# Discord Bot
discord.py>=2.3.0

# Google Gemini AI
google-generativeai>=0.3.0

# GitHub Integration
PyGithub>=2.1.1

# Environment Variables
python-dotenv>=1.0.0

# Configuration Management
pyyaml>=6.0

# Data Validation
pydantic>=2.0.0

# Async HTTP
aiohttp>=3.9.0

# Date/Time Handling
python-dateutil>=2.8.0

# Task Scheduling
APScheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
